# Performance improvement of remove_unnatural()