I fixed the performance of remove_unnatural().
What I did is as following:
 - Removed casting a amino_acids into a set in the main loop
 - Escaped from the main loop instantly when unnatural letter found

I measured performance using 10000 seqs of length between 10 to 500. About 1000 seqs are unnatural in the dataset.

Original: 2.45 sec
Stop casting: 1.60 sec
Escape: 1.09 sec
Both: 0.27 sec